Pass TranslationBlock to tb_exec for TCG
Standard QEMU passes only the pointer to the start of the translation
block's instructions, however for Qira tracking we need more details
from the `TranslationBlock` structure.

This updates the `tcg_qemu_tb_exec` function to take a pointer to the
`TranslationBlock` to execute, to minimize the changes needed to
build with the Qira patch.

We should probably determine if it would be possible to move the Qira
tracking code into `cpu_tb_exec` instead, which would not require
modifying the signature of `tcg_qemu_tb_exec` (although it is only
called from that one location).